rename ops module to bin_op;
