+++ ergo-lib/src/ast.rs
@@ -1,11 +1,11 @@
 //! AST for ErgoTree
 
+pub(crate) mod bin_op;
 pub(crate) mod coll_fold;
 pub(crate) mod expr;
 pub(crate) mod extract_reg_as;
 pub(crate) mod global_vars;
 pub(crate) mod method_call;
-pub(crate) mod ops;
 pub(crate) mod option_get;
 pub(crate) mod predef_func;
 pub(crate) mod property_call;
